+++ landingzones/landingzone_secure_vnet_dmz/net_core/blueprint.tf
@@ -18,10 +18,10 @@ resource "azurerm_resource_group" "rg_edge" {
 
 module "core_network" {
   source  = "aztfmod/caf-virtual-network/azurerm"
-  version = "1.1.0"
+  version = "2.0.0"
 
   convention                        = var.global_settings.convention
-  virtual_network_rg                = azurerm_resource_group.rg_network.name
+  resource_group_name               = azurerm_resource_group.rg_network.name
   prefix                            = var.prefix
   location                          = var.global_settings.location_map.region1
   networking_object                 = var.core_networking.shared_services_vnet
@@ -33,12 +33,12 @@ module "core_network" {
 
 module "az_firewall_ip" {
   source  = "aztfmod/caf-public-ip/azurerm"
-  version = "1.0.0"
+  version = "2.0.0"
 
   convention                       = var.global_settings.convention 
   name                             = var.core_networking.ip_addr_config.ip_name
   location                         = var.location
-  rg                               = azurerm_resource_group.rg_edge.name
+  resource_group_name              = azurerm_resource_group.rg_edge.name
   ip_addr                          = var.core_networking.ip_addr_config
   tags                             = var.global_settings.tags_hub
   diagnostics_map                  = var.caf_foundations_accounting.diagnostics_map
@@ -48,11 +48,11 @@ module "az_firewall_ip" {
 
 module "az_firewall" {
   source  = "aztfmod/caf-azure-firewall/azurerm"
-  version = "1.1.0"
+  version = "2.0.0"
 
   convention                        = var.global_settings.convention 
   name                              = var.core_networking.az_fw_config.name
-  rg                                = azurerm_resource_group.rg_network.name
+  resource_group_name               = azurerm_resource_group.rg_network.name
   subnet_id                         = lookup(module.core_network.vnet_subnets, "AzureFirewallSubnet", null)
   public_ip_id                      = module.az_firewall_ip.id
   location                          = var.global_settings.location_map.region1
@@ -135,12 +135,12 @@ module "user_route_transit_to_az_firewall" {
 ## VPN Gateway
 module "vpn_pip" {
   source  = "aztfmod/caf-public-ip/azurerm"
-  version = "1.0.0"
+  version = "2.0.0"
 
   convention                       = var.global_settings.convention 
   name                             = var.core_networking.gateway_config.pip.name
   location                         = var.location
-  rg                               = azurerm_resource_group.rg_transit.name
+  resource_group_name              = azurerm_resource_group.rg_transit.name
   ip_addr                          = var.core_networking.gateway_config.pip
   tags                             = var.global_settings.tags_hub
   diagnostics_map                  = var.caf_foundations_accounting.diagnostics_map